third_party/WebKit/LayoutTests/external/wpt/css/css-position/position-sticky-transforms.html

<!DOCTYPE html>
<title>transforms on position:sticky elements should apply after sticking</title>
<link rel="help" href="https://www.w3.org/TR/css-position-3/#sticky-pos" />
<meta name="assert" content="This test checks that transforms on position:sticky elements are carried out on their stuck position" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script src="resources/sticky-util.js"></script>

<body></body>

<style>
.perspective {
  transform: perspective(3px) translateZ(1px);
  height: 50px;
  width: 50px;
}
</style>

<script>
test(() => {
  const elements = setupStickyTest('top', 50);
  elements.sticky.style.transform = 'scale(2)';
  elements.scroller.scrollTop = 200;

  // Transforms don't affect offsetTop, so use getBoundingClientRect.
  // Scaling the sticky element by 2 means its top-y moves (1/2 * height)
  // upwards, in this case placing it at the top of the viewport.
  const boundingRect = elements.sticky.getBoundingClientRect();
  assert_equals(boundingRect.y, elements.scroller.getBoundingClientRect().y);
}, 'test');

test(() => {
  const elements = setupStickyTest('top', 50);
  elements.sticky.style.transform = 'rotateX(60deg)';
  elements.scroller.scrollTop = 200;
}, 'test2');

/*
window.addEventListener('load', function() {
  document.getElementById('scroller1').scrollTop = 50;
  document.getElementById('scroller2').scrollTop = 50;
  document.getElementById('scroller3').scrollTop = 50;
});
*/
</script>
